Rename the orderItemsTable component
